refactor : 채널 상세 페이지 UI 구조화
