+++ lib/presentation/pages/youtube/channel_detail/channel_detail_state.dart
@@ -1,8 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
-import 'package:techtalk/features/youtube/data_source/remote/models/youtube_content_overview_model.dart';
-import 'package:techtalk/features/youtube/data_source/remote/models/youtube_main_model.dart';
+import 'package:techtalk/features/youtube/index.dart';
 import 'package:techtalk/features/youtube/repositories/entities/channel_detail_entity.dart';
 import 'package:techtalk/presentation/pages/youtube/channel_detail/provider/channel_contents_pagination_provider.dart';
 import 'package:techtalk/presentation/pages/youtube/channel_detail/provider/channel_detail_provider.dart';
@@ -21,8 +20,15 @@ mixin class ChannelDetailState {
   }
 
   ///
+  /// 채널 정보
+  ///
+  ChannelEntity channel(WidgetRef ref) =>
+      ref.read(channelDetailRouteArgProvider);
+
+  ///
   /// 채널 상세 정보
   ///
+  @Deprecated('현재 YoutubeExplore 엔드포인트가 안맞아서 상세 정보를 못불러오고 있음')
   AsyncValue<ChannelDetailEntity> channelDetail(WidgetRef ref) {
     final channelId = ref.read(channelDetailRouteArgProvider).id;
     return ref.watch(channelDetailProvider(channelId));
