+++ lib/presentation/pages/youtube/channel_detail/widgets/channel_info_view.p.dart
@@ -1,0 +1,27 @@
+part of '../channel_detail_page.dart';
+
+class _ChannelInfoView extends ConsumerWidget with ChannelDetailState {
+  const _ChannelInfoView({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return Row(
+      children: [
+        RoundProfileImg(
+          size: 64,
+          imgUrl: channel(ref).logoUrl,
+        ),
+        const Gap(12),
+        Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Text(
+              channel(ref).name,
+              style: AppTextStyle.headline2,
+            ),
+          ],
+        ),
+      ],
+    );
+  }
+}
